requirements.txt
langchain
langgraph
langchain-ollama
langchain-qdrant
langchain-community
qdrant-client
pymupdf>=1.23.0
tabulate
streamlit
plotly
pandas
python-dotenv
